requirements.txt
# requirements.txt - Comprehensive dependencies for both Reading App and YouTube Quiz

python-dotenv

# Reading App Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
python-docx>=0.8.11

# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0

# AI/LLM Dependencies (shared)
google-generativeai>=0.6
langchain>=0.2
langchain-google-genai>=1.0
openai

# Text Processing
nltk>=3.8
spacy>=3.6.0

# Audio Processing (for Reading App)
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.0

# Utilities
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0

# Optional
torch>=2.0.0
transformers>=4.30.0

# Development/Testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0